requirements.txt
aiodns==3.0.0
aiofile==3.8.1
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
blinker==1.5
blobfile==2.0.1
bs4==0.0.1
cachetools==5.2.0
caio==0.9.11
camelot-py==0.10.1
certifi==2021.10.8
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.0.9
click==8.0.3
commonmark==0.9.1
cryptography==39.0.0
dataclasses-json==0.5.7
decorator==4.4.2
docx==0.2.4
docx2txt==0.8
entrypoints==0.4
et-xmlfile==1.1.0
faiss-cpu==1.7.3
filelock==3.9.0
Flask==2.2.3
Flask-Cors==3.0.10
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.30
gunicorn==20.1.0
idna==3.3
imageio==2.26.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.0.0
install==1.3.5
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.2.0
jsonschema==4.17.3
langchain==0.0.107
llama-index==0.4.26
lxml==4.9.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.0
numpy==1.21.4
openai==0.27.0
opencv-python==4.7.0.68
openpyxl==3.0.9
packaging==22.0
pandas==1.3.4
pandas-stubs==1.2.0.39
pdfminer.six==20221105
pdfplumber==0.7.6
Pillow==9.4.0
proglog==0.1.10
protobuf==3.20.3
pyarrow==10.0.1
pycares==4.3.0
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.6
pydeck==0.8.0
Pygments==2.14.0
PyJWT==2.6.0
Pympler==1.0.1
pypdf==3.5.1
PyPDF2==2.12.1
PyQt3D==5.15.5
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rich==13.0.0
scikit-learn==1.2.2
scipy==1.10.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
soupsieve==2.4
SQLAlchemy==1.4.46
stream-chat==4.6.0
streamlit==1.20.0
streamlit-chat==0.0.2.2
streamlit-ext==0.1.6
sv-ttk==2.4
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.0
toml==0.10.2
toolz==0.12.0
torch==2.0.0
tornado==6.2
tqdm==4.62.3
typing==3.7.4.3
typing-inspect==0.8.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.7
validators==0.20.0
Wand==0.6.10
Werkzeug==2.2.3
whisper==1.1.10
yarl==1.8.2
zipp==3.11.0
